make the 'cached' flag per-phase
This allows to recognize that a test was executed while the package
prefix was pulled from cache. It wasn't meant to be a normal mode of
execution, but it turns out that because of external events (e.g.
buildbot being turned off, the VM being evicted), we have to deal
with the build passing but tests not having been executed.