force-push packages that have been rebuilt
This is a bugfix for a missing cache update if a package has been
built in a run, but its test (and, later, other utilities) were
executed later.

In this scenario, `push_package_to_cache` would not update the
metadata, which would cause the next CI build to rebuild the
package ... and forever so.

By force-pushing a package if it was built locally, we fix this
problem. This condition works because we consider that a package
on which we need to run some utilities needs to be rebuilt (see
https://github.com/rock-core/autoproj-ci/blob/37df655ec1827de9ad7d0da49e66d16a8675d811/lib/autoproj/cli/ci.rb#L216)